requirements.txt
bibtexparser==1.2.0
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.6.0
control==0.8.3
cvxopt==1.2.5
cycler==0.10.0
Cython==0.29.21
dill==0.3.2
docopt==0.6.2
future==0.18.2
graphviz==0.17
gym==0.17.3
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
mpmath==1.1.0
numpy==1.19.2
Pillow==8.0.0
pipreqs==0.4.10
progressbar2==3.53.1
pybullet==3.0.6
pygame==2.0.1
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
PyYAML==5.3.1
qpsolvers==1.4
quadprog==0.1.7
rebiber==1.1.1
requests==2.24.0
scipy==1.5.3
six==1.15.0
sympy==1.6.2
tqdm==4.57.0
urllib3==1.25.11
yarg==0.1.9
